[#261] BC Break: move psr7 handler factories to psr7 package
