+++ packages/toggle-crud-psr7-api/src/DeleteFeatureFactory.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Pheature\Crud\Psr11\Toggle;
+namespace Pheature\Crud\Psr7\Toggle;
 
 use Pheature\Crud\Psr7\Toggle\DeleteFeature;
 use Pheature\Crud\Toggle\Handler\RemoveFeature;
