+++ packages/toggle-crud-psr7-api/src/GetFeaturesFactory.php
@@ -2,16 +2,16 @@
 
 declare(strict_types=1);
 
-namespace Pheature\Crud\Psr11\Toggle;
+namespace Pheature\Crud\Psr7\Toggle;
 
 use Pheature\Core\Toggle\Read\FeatureFinder;
-use Pheature\Crud\Psr7\Toggle\GetFeature;
+use Pheature\Crud\Psr7\Toggle\GetFeatures;
 use Psr\Container\ContainerInterface;
 use Psr\Http\Message\ResponseFactoryInterface;
 
-final class GetFeatureFactory
+final class GetFeaturesFactory
 {
-    public function __invoke(ContainerInterface $container): GetFeature
+    public function __invoke(ContainerInterface $container): GetFeatures
     {
         /** @var FeatureFinder $featureFinder */
         $featureFinder = $container->get(FeatureFinder::class);
@@ -21,8 +21,8 @@ final class GetFeatureFactory
         return self::create($featureFinder, $responseFactory);
     }
 
-    public static function create(FeatureFinder $featureFinder, ResponseFactoryInterface $responseFactory): GetFeature
+    public static function create(FeatureFinder $featureFinder, ResponseFactoryInterface $responseFactory): GetFeatures
     {
-        return new GetFeature($featureFinder, $responseFactory);
+        return new GetFeatures($featureFinder, $responseFactory);
     }
 }
