+++ packages/toggle-crud-psr7-api/src/PostFeatureFactory.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Pheature\Crud\Psr11\Toggle;
+namespace Pheature\Crud\Psr7\Toggle;
 
 use Pheature\Crud\Psr7\Toggle\PostFeature;
 use Pheature\Crud\Toggle\Handler\CreateFeature;
